requirements.txt
Django>=1.3
python-slugify>=0.0.2

